requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
billiard==3.6.4.0
boto3==1.20.7
botocore==1.23.7
celery==5.2.7
certifi==2021.10.8
cffi==1.15.0
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==38.0.4
daphne==4.0.0
Django==3.2.9
django-celery-beat==2.4.0
django-cors-headers==3.10.0
django-environ==0.8.1
django-filter==23.2
django-multiselectfield==0.1.12
django-o365mail==1.0.1
django-rest-knox==4.1.0
django-rest-passwordreset==1.2.1
django-timezone-field==5.0
djangorestframework==3.12.4
exceptiongroup==1.1.1
gevent==22.10.2
greenlet==2.0.1
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.3
incremental==22.10.0
iniconfig==2.0.0
isort==5.10.1
jmespath==0.10.0
jsonpath==0.82
kombu==5.2.4
lazy-object-proxy==1.7.1
mccabe==0.6.1
msgpack==1.0.4
msgpack-python==0.5.6
mysqlclient==2.1.1
numpy==1.23.5
O365==2.0.19
oauthlib==3.2.0
opencv-python==4.6.0.66
packaging==23.1
Pillow==9.4.0
platformdirs==2.5.1
pluggy==1.0.0
prompt-toolkit==3.0.36
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.20
pyOpenSSL==22.1.0
pytest==7.3.1
pytest-django==4.5.2
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
redis==4.4.2
requests==2.27.1
requests-oauthlib==1.3.1
s3transfer==0.5.0
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stringcase==1.2.0
stripe==4.0.2
toml==0.10.2
tomli==2.0.1
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
websocket==0.2.1
websocket-client==1.6.4
wrapt==1.15.0
zope.event==4.6
zope.interface==5.5.2
